+++ src/flow/platform/attributes.js
@@ -1,0 +1,6 @@
+// @flow
+
+import type { ReactNativeStyleAttributeName } from './attributes-native'
+import type { CSSPropertyName } from './attributes-web'
+
+export type StyleAttributeName = CSSPropertyName | ReactNativeStyleAttributeName