+++ src/flow/platform/style.js
@@ -1,0 +1,6 @@
+// @flow
+
+export type ReactNativeStyleObject = { [string]: mixed }
+export type CSSClassName = string
+
+export type PredefinedStyle = ReactNativeStyleObject | CSSClassName