+++ src/flow/platform/stylesheetOutput.js
@@ -1,0 +1,5 @@
+// @flow
+import type { PredefinedStyle } from './style'
+
+// NOTE: This file is used when user doesn't have .web.js, .native.js imports configured in flow
+export type ZacsStylesheet = { [string]: PredefinedStyle }