+++ src/flow/platform/stylesheetOutput.web.js
@@ -1,0 +1,4 @@
+// @flow
+import type { CSSClassName } from './style'
+
+export type ZacsStylesheet = { [string]: CSSClassName }