+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file.
 
+## Unreleased
+
+- BREAKING: If you use Flow, you need to upgrade to >=0.185.2
+- Officially release `zacs.stylesheet({})`
+- Flow typing improvements - `zacs.styled()` now returns correctly typed React component in most cases
+
 ## 1.1.0 (2020-10-09)
 
 - You can now declare multiple unconditional styles by passing an array of stylesets/class names